src/requirements.txt
Django==4.2.5
djoser==2.2.0
djangorestframework==3.14.0
django-redis==5.4.0
psycopg2==2.9.7
celery[redis]==5.3.4
python-binance==1.0.19
tinkoff-investments==0.2.0b60
python-dotenv==1.0.0
flower==2.0.1
django_celery_beat==2.5.0
scipy==1.11.4
